// src/types/blog.ts
export interface BlogContent {
  title: string;
  excerpt: string;
  content: string;
}

export interface BlogPost {
  id: string;
  slug: string;
  coverImage: string;
  tags: string[];
  author: string;
  date: string;
  readTime: string;
  published: boolean;
  content: Record<string, BlogContent>;
}

export type CreateBlogPostDto = Omit<BlogPost, 'id'>;
export type UpdateBlogPostDto = Partial<CreateBlogPostDto>;